move app from git to separate module
