+++ jhub_apps/service/routes.py
@@ -35,8 +35,8 @@ from jhub_apps.service.utils import (
     get_jupyterhub_config,
     get_spawner_profiles,
     get_thumbnail_data_url,
-    get_shared_servers, get_app_configuration_from_git,
-)
+    get_shared_servers, )
+from jhub_apps.service.app_from_git import get_app_configuration_from_git
 from jhub_apps.spawner.types import FRAMEWORKS
 from jhub_apps.version import get_version
